Fix #501 need commas after deps move random_base62 to util, deprecate pkio
